refactor: Apply MVVM Pattern
- Package를 Model, View, ViewModel로 나누어서 보기 편하게 하였다.
