+++ app/src/main/java/campus/tech/kakao/map/model/LocationDbHelper.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.db
+package campus.tech.kakao.map.model
 
 import android.content.ContentValues
 import android.content.Context
