+++ app/src/main/java/campus/tech/kakao/map/model/SearchLocationRepository.kt
@@ -1,0 +1,5 @@
+package campus.tech.kakao.map.model
+
+class SearchLocationRepository {
+
+}
