+++ app/src/main/java/campus/tech/kakao/map/viewmodel/SearchLocationViewModel.kt
@@ -1,0 +1,7 @@
+package campus.tech.kakao.map.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class SearchLocationViewModel : ViewModel() {
+
+}
