+++ app/src/main/res/layout/activity_search_location.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".SearchLocationActivity">
+    tools:context=".view.SearchLocationActivity">
 
     <EditText
         android:id="@+id/searchInputEditText"
